+++ packages/zosfiles/src/cli/hMigrate/HMigrate.definition.ts
@@ -18,7 +18,7 @@ import i18nTypings from "../-strings-/en";
 const { DESCRIPTION } = (require("../-strings-/en").default as typeof i18nTypings).HMIGRATE;
 
 /**
- * Download group definition containing its description and children
+ * hMigrate group definition containing its description and children
  * @type {ICommandDefinition}
  */
 export const HMigrateDefinition: ICommandDefinition = {