+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+
+- Fix compilation error
+
+## `6.19.0`
+
+- Add CLI command to delete migrated data sets `zowe zos-files delete migrated-data-sets`.
+
 ## `6.18.0`
 
 - Add the --fail-fast option to the `zowe zos-files download all-members` command